requirements.txt
# Core dependencies
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
timm>=0.9.10

# Image processing and augmentations
albumentations>=1.3.1
opencv-python>=4.8.1
Pillow>=10.0.1

# Data handling
numpy>=1.24.0
pandas>=2.1.0
h5py>=3.10.0

# Configuration and validation
pydantic>=2.4.0
omegaconf>=2.3.0
PyYAML>=6.0.1

# Experiment tracking
wandb>=0.16.0
mlflow>=2.8.0
tensorboard>=2.15.0

# API framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Scientific computing
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Distributed training
accelerate>=0.24.0

# Model optimization
onnx>=1.15.0
onnxruntime>=1.16.0

# Monitoring
prometheus-client>=0.19.0
psutil>=5.9.0

# Utilities
tqdm>=4.66.0
rich>=13.6.0
typer>=0.9.0